Make test_scaffold_assets_with_source() order-independent
Previously, this test was order-dependent (re-ordering the elements of mc.model_parameters.data would fail the test).

However, mc.model_parameters.data's ordering is not guaranteed. It should be treated more like a set than a sequence.

This commit rewrites the test to pass for both orderings ([train, eval] and [eval, train]).

PiperOrigin-RevId: 409194067