don't raise ValueError for non-handled metric types
Today, when MCT encounters a non-handled metric type, it raises an error and terminates. This makes MCT unusable for any user with non-handled metric types.

Instead of raising an error, we should log a warning so that users are aware MCT does not handle their metric type, but does not block them from using MCT altogether.

PiperOrigin-RevId: 415523358