add `artifact_type_id` arg
## Why we need this

TFX Executors do not read a MLMD store directly. Instead, their input/output API consists of MLMD artifacts. So we have access to the ModelCard Artifact and all its properties, but not the MLMD store itself.

Previously, ModelCardToolkit required a MLMD store to generate an artifact. This means that ModelCardToolkit would not work as an Executor.

## What this does (API)

Introduces a new arg to ModelCardToolkit, `artifact_type_id`. This is the id for the ModelCard Artifact type registered to some MLMD store.

This is only used when ModelCardToolkit is used as part of Executor.

## What this does (internal logic)

Decouples MLMD saving logic from artifact creation logic, so artifacts can be created even when a MLMD store doesn't exist.

This also gets rid of the `artifact_id` property and replaces it with the `artifact` property. This is because `artifact_id` is only set once the artifact has been saved to MLMD. Since we no longer assume all artifacts are saved to MLMD, it no longer makes sense to refer to artifacts by their id.

PiperOrigin-RevId: 421850889