Decouple MCT core logic from artifact generation logic
Previously, ModelCardToolkit was responsible for creating both the Model Card assets (proto, json, html, md, etc.) and the ModelCard artifact (a TFX/MLMD artifact which stores a reference to those assets.

This tight coupling is bad design; these two responsibilities should be separated into independent modules (which this commit accomplishes).

In addition to improving design, this change is also a prerequisite for building the TFX Executor. The Executor pattern takes an artifact as input, writes assets to the artifact, and returns the artifact. MCT's previous behavior coupled creation of the artifact with creation of the assets, which was incompatible with the Executor pattern.

PiperOrigin-RevId: 422667916