Model Card Toolkit Executor
This is the Executor class for the Model Card TFX Component. It reads TFMA, TFDV, and PushedModel artifacts, and returns a ModelCard artifact.

This commit also changes `source` to use the generic `tfx.types.Artifact` instead of the specific `standard_artifacts`. This is needed because Python type checking does not know the type of artifacts in `input_dict`, so we need to use the generic type instead.

PiperOrigin-RevId: 423072443